Fixed javascript loading. Params are appended to urls when submitting forms though.
'application.css' isn't loaded automatically for some reason, so it needs to be explicitly included.
'prototype.js' is a javascript framework akin to jQuery that should eventually be removed.
'jquery' is included in with rails by default as of Rails 3.1.
Can't find where default javascripts are set and it doesn't seem to affect the main page or signup/login forms.